backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.4,<5.5
django-cors-headers>=4.4,<4.5
django-filter>=24.0
psycopg2-binary>=2.9
channels>=4.3,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0
Pillow>=10.0
